requirements.txt
Flask>=2.0.0
plotly>=5.0.0
requests>=2.25.0
python-dotenv>=0.19.0
google-generativeai>=0.2.0